fix: void `Xsession` fail when $HOME is full
When user home is full,
scripts write to a file in user home will failed.

After redirect stdout and stderr to ERRFILE,
aka $HOME/.xsession-errors,
this script echo a line to that file.

This echo might failed and make this script exit with error.
In that case, the session leader will not able to even start.

Releated: linuxdeepin/developer-center#4685
Releated: linuxdeepin/dde-daemon#526
Signed-off-by: black-desk <me@black-desk.cn>